Breaking down UI plugins into smaller pieces
