+++ sbt-core-next/src/main/scala/sbt/plugins/CommandLineUIServices.scala
@@ -1,0 +1,24 @@
+package sbt
+package plugins
+
+import sbt.serialization._
+
+private[sbt] object CommandLineUIServices extends SbtPrivateInteractionService with SbtPrivateSendEventService {
+  override def readLine(prompt: String, mask: Boolean): Option[String] = {
+    val maskChar = if (mask) Some('*') else None
+    SimpleReader.readLine(prompt, maskChar)
+  }
+  // TODO - Implement this better!
+  def confirm(msg: String): Boolean = {
+    object Assent {
+      def unapply(in: String): Boolean = {
+        (in == "y" || in == "yes")
+      }
+    }
+    SimpleReader.readLine(msg + " (yes/no): ", None) match {
+      case Some(Assent()) => true
+      case _ => false
+    }
+  }
+  override def sendEvent[T: Pickler](event: T): Unit = ()
+}
